Separating playing with registration in rspec, moved helper methods where they actually belong. More to come soon.
