+++ spec/spec_helper.rb
@@ -118,3 +118,8 @@ SimpleCov.start
 require 'byebug'
 
 Capybara.app = Rps
+
+
+def current_path_info
+  current_url.sub(%r{.*?://},'')[%r{[/\?\#].*}] || '/'
+end
