+++ spec/web_helpers.rb
@@ -1,0 +1,6 @@
+require 'selenium-webdriver'
+
+def browser_view
+  include Capybara::DSL
+  Capybara.default_driver = :selenium
+end
